+++ src/util.hpp
@@ -15,7 +15,7 @@ namespace util {
 	const glm::vec3 up = glm::vec3(0,1,0);
 
 	glm::vec3 std2glm(std::vector<double>);
-	std::vector<float> glm2std(glm::vec3 v);
+	std::vector<double> glm2std(glm::vec3 v);
 
 	std::vector<geometry> load_scene_full_mesh(const char* filename, bool smooth);
 	glm::vec3 gs1(glm::vec3 a, glm::vec3 b);
